256 pixel width does NOT fit in uint8_t...

+++ source_code/src/lib/oledmp/bitmap.h
@@ -24,7 +24,7 @@
 #define BITMAP_H_
 
 typedef struct bitmap_s {
-    uint8_t width;	//*< width of image in pixels
+    uint16_t width;	//*< width of image in pixels
     uint8_t height;	//*< height of image in pixels
     uint8_t depth;	//*< Number of bits per pixel
     uint16_t dataSize;  //*< number of words in data
